wrote checkWin method, and added the cheeckWin method in to the while loop with the 2 expected params
then tried to test code, but as testing it, there was an issue with CheckDiagonalLines2 method, after looking at this I found that the if statement I wrote was incorrect. Looking in to this and experimenting, I have now amended this.

Tested the code again, and after each turn, it was populating the correct position, but didn’t recognise the win  line?